API_REST/requirements.txt
django
djangorestframework
gunicorn
mysqlclient
coverage
django-heroku
django-cors-headers
